+++ printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: rprieto- <rprieto-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 12:42:43 by rprieto-          #+#    #+#             */
-/*   Updated: 2020/10/15 16:02:02 by rprieto-         ###   ########.fr       */
+/*   Updated: 2020/10/19 00:30:35 by rprieto-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,5 +71,5 @@ void				ft_lstiter(t_list *lst, void (*f)(void *));
 t_bool				ft_checkchar(char c, char *set);
 int					ft_get_index_of(char *string, char c);
 int					ft_nbrlen(long n);
-int					ft_abs(int n);
+long				ft_abs(long n);
 #endif
